feat:添加流媒体接口

+++ models/models.go
@@ -37,14 +37,14 @@ type Dir struct {
 
 // 歌曲
 type Song struct {
-	Title  string  `gorm:"not null" json:"title"`          // 标题
-	Artist string  `json:"artist"`                         // 作者
-	Album  string  `json:"album"`                          // 专辑
-	Type   string  `gorm:"not null" json:"type"`           // 文件类型
-	Size   float64 `grom:"not null" json:"size"`           // 大小
-	Dir    int     `gorm:"not null" json:"dir"`            // 目录
-	Time   int     `gorm:"not null,default=0" json:"time"` // 时长
-	Path   string  `gorm:"not null,default=/" json:"path"` // 位置url
+	Title  string `gorm:"not null" json:"title"`          // 标题
+	Artist string `json:"artist"`                         // 作者
+	Album  string `json:"album"`                          // 专辑
+	Type   string `gorm:"not null" json:"type"`           // 文件类型
+	Size   int64  `grom:"not null" json:"size"`           // 大小
+	Dir    int    `gorm:"not null" json:"dir"`            // 目录
+	Time   int    `gorm:"not null,default=0" json:"time"` // 时长
+	Path   string `gorm:"not null,default=/" json:"path"` // 位置url
 
 	Model
 }
